Handle nssm parameter diff & quoting
Recent version of nssm now better handle command line parameters.
Passing arguments properly quoted is easier.

This commit quotes & escape parameters in a way that the dump output
is the same as the input. It helps for idempotency and ease usage of
the resource.

No need to triple quotes arguments :)
But this is a breaking change because it changes the public interface
of this resource, even if the behavior was already broken following
last nssm bump.